% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drain_dem.R
\name{drain_dem}
\alias{drain_dem}
\title{Drain Your Dammed DEM}
\usage{
drain_dem(x, label, ...)
}
\arguments{
\item{x}{A \link[raster:raster]{raster::raster} DEM. Alternatively, a \link[sf:st_bbox]{sf::st_bbox} object for which a DEM will
be downloaded}

\item{label}{A character string naming the study area.}

\item{...}{Arguments passed on to various \code{FedData} functions.}
}
\value{
A drained \link[raster:raster]{raster::raster} DEM.
}
\description{
Given a digital elevation model (DEM) with human-made dammed reservoirs,
returns a version with estimates for elevations under reservoirs and dams.
}
\examples{
library(magrittr)
library(demdrainer)

dem <-
  FedData::get_ned(demdrainer::mcphee,
    label = "McPhee"
  )

drained_dem <-
  drain_dem(dem,
    label = "McPhee"
  )

raster::plot(dem)
raster::plot(drained_dem)
}
