% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tidy-eval.R
\docType{data}
\name{mcphee}
\alias{mcphee}
\title{Polygon of bounding box around McPhee Reservoir, Colorado}
\format{
A geometry set with one feature.
}
\usage{
mcphee
}
\description{
A dataset containing the bounding box around McPhee Reservoir, Colorado.
}
\keyword{datasets}
